Fix type bug in copyfits CP_INIT, fix bug in compositing
In copyfits float images were only half-initialised due
to sizeof(WORD) being used instead of sizeof(float)

In compositing the first image to be loaded is copied to
gfit using CP_EXPAND if it is single channel, but it
was missing CP_INIT to ensure the additional channels
are initialized.

+++ src/io/image_format_fits.c
@@ -2561,7 +2561,7 @@ int copyfits(fits *from, fits *to, unsigned char oper, int layer) {
 
 			if ((oper & CP_INIT)) {
 				// clearing to->fdata allocated above
-				memset(to->fdata, 0, nbdata * depth * sizeof(WORD));
+				memset(to->fdata, 0, nbdata * depth * sizeof(float));
 			}
 		}
 		else {
